% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{algorithm_cluster}
\alias{algorithm_cluster}
\title{cluster package algorithms}
\usage{
algorithm_cluster()
}
\value{
list with the algorithms
}
\description{
cluster package algorithms
}
\keyword{internal}
