% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{algorithm_clusterr}
\alias{algorithm_clusterr}
\title{ClusterR package algorithms}
\usage{
algorithm_clusterr()
}
\value{
list with the algorithms
}
\description{
ClusterR package algorithms
}
\keyword{internal}
