% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{algorithms}
\alias{algorithms}
\title{Method that returns the list of used algorithms}
\usage{
algorithms()
}
\value{
algorithm listing array
}
\description{
Method that returns the list of used algorithms
}
\keyword{internal}
