% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{algorithms_package}
\alias{algorithms_package}
\title{Method that returns all the algorithms executed by the package}
\usage{
algorithms_package(packages)
}
\arguments{
\item{packages}{package array}
}
\value{
array with the algorithms we're going to run
}
\description{
Method that returns all the algorithms executed by the package
}
\keyword{internal}
