% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{appClustering}
\alias{appClustering}
\title{Method that allows us to execute the main algorithm in graphic interface mode instead of through the console.}
\usage{
appClustering()
}
\description{
Method that allows us to execute the main algorithm in graphic interface mode instead of through the console.
}
\examples{

appClustering()

}
