% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball-data.R
\docType{data}
\name{basketball}
\alias{basketball}
\title{This data set contains a series of statistics (5 attributes) about 96 basketball players:}
\format{
A data frame with 96 observations on 5 variables:
\describe{
  This data set contains a series of statistics about basketball players:
  \item{assists_per_minuteReal}{average number of assistances per minute}
  \item{heightInteger}{height of the player}
  \item{time_playedReal}{time played by the player}
  \item{ageInteger}{number of years of the player}
  \item{points_per_minuteReal}{average number of points per minute}
}
}
\source{
KEEL, <http://www.keel.es/>
}
\usage{
data(basketball)
}
\description{
This data set contains a series of statistics about basketball players:
}
\keyword{datasets}
