% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{best_ranked_external_metrics}
\alias{best_ranked_external_metrics}
\title{Best rated external metrics.}
\usage{
best_ranked_external_metrics(df)
}
\arguments{
\item{df}{matrix or data frame with the result of running the clustering algorithm.}
}
\value{
Returns a data.frame with the best classified external variables.
}
\description{
method in charge of searching for each algorithm those that have the best external classification.

Method that looks for those external variables that are better classified,
making use of the ranking column. In this way we discard the rest of the variables and
only work with those that give the best response to the algorithm in question.
}
\examples{

result = clustering(
               df = cluster::agriculture,
               min = 4,
               max = 5,
               algorithm='gmm',
               metrics=c("recall"),
               variables = FALSE
         )

best_ranked_external_metrics(df = result)

\dontrun{
best_ranked_external_metrics(df = result$result)
}

}
