% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{best_ranked_internal_metrics}
\alias{best_ranked_internal_metrics}
\title{Method that calculates the best rated internal metrics.}
\usage{
best_ranked_internal_metrics(df)
}
\arguments{
\item{df}{data matrix or data frame with the result of running the clustering algorithm.}
}
\value{
returns a table with the internal metrics that has the best rating.
}
\description{
Method that calculates the best rated internal metrics.
}
\examples{

result = clustering(
               df = cluster::agriculture,
               min = 4,
               max = 5,
               algorithm='gmm',
               metrics=c("dunn"),
               variables = FALSE
         )

best_ranked_internal_metrics(result)

}
