% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calculate_best_external_variables_by_metrics}
\alias{calculate_best_external_variables_by_metrics}
\title{Method that calculates the best rated external metrics.}
\usage{
calculate_best_external_variables_by_metrics(df)
}
\arguments{
\item{df}{Data matrix or data frame.}
}
\value{
Return a table with the external metrics that has the best rating.
}
\description{
Method that calculates the best rated external metrics.
}
\keyword{internal}
