% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calculate_best_validation_internal_by_metrics}
\alias{calculate_best_validation_internal_by_metrics}
\title{Method that calculates which algorithm and which metric behaves best for the datasets provided}
\usage{
calculate_best_validation_internal_by_metrics(df)
}
\arguments{
\item{df}{data matrix or data frame}
}
\value{
returns a table with the algorithm and the best performing metric for the datasets
}
\description{
Method that calculates which algorithm and which metric behaves best for the datasets provided
}
\keyword{internal}
