% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validation.R
\name{calculate_connectivity}
\alias{calculate_connectivity}
\title{Method to calculate the Connectivity}
\usage{
calculate_connectivity(
  distance = NULL,
  clusters,
  datadf = NULL,
  neighbSize = 12,
  method = "euclidean"
)
}
\arguments{
\item{distance}{Dissimilarity matrix.}

\item{clusters}{Array that containe tha data grouped in cluster.}

\item{neighbSize}{Number of neighbours.}

\item{method}{Indicates the method for calculating distance between points.
Default is euclidean.}

\item{datdf}{Dataframe with original data.}
}
\value{
Return a double with the result of the connectivity calculation.
}
\description{
Method to calculate the Connectivity
}
\keyword{internal}
