% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validation.R
\name{calculate_dunn}
\alias{calculate_dunn}
\title{Method to calculate the dunn.}
\usage{
calculate_dunn(distance = NULL, clusters, datadf = NULL, method = "euclidean")
}
\arguments{
\item{distance}{Dissimilarity matrix.}

\item{clusters}{Array that containe tha data grouped in cluster.}

\item{datadf}{Dataframe with original data.}

\item{method}{Indicate the method for calculating distance between points.}
}
\value{
Return a double with the result of the dunn calculation
}
\description{
Method to calculate the dunn.
}
\keyword{internal}
