% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{calculate_k_projection}
\alias{calculate_k_projection}
\title{K-Projection.}
\usage{
calculate_k_projection(data, dim_k, arg_kernel)
}
\arguments{
\item{data}{matrix, with columns of numeric data.
Each row represents one data point. The columns are the dimensions.}

\item{dim_k}{numeric value smaller than the number of data points, the dimension the data should be projected into.}

\item{arg_kernel}{Optional arguments for the \code{calculate_mercer_kernel} function.

TODO}
}
\value{
matrix, with the transformed data points
}
\description{
\code{calculate_k_projection} projects the data into the vector space R^\code{dim_k}
}
\section{terms}{

All data points in the input data have to stem from an identical distribution and have to be independent.

Clearly the data points also have to be numeric and from the same vector space.
}

\references{
\url{https://link.springer.com/book/10.1007/978-3-662-59354-7}
}
