% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calculate_result_internal}
\alias{calculate_result_internal}
\title{Method that returns the value or variable depending on where it is in the
calculated metrics.}
\usage{
calculate_result_internal(
  algorith,
  distance,
  cluster,
  dataset,
  ranking,
  timeInternal,
  dunn,
  connectivity,
  silhouette,
  variables
)
}
\arguments{
\item{algorith}{Algorithm name.}

\item{distance}{Name of the metric used to calculate the distance between
points.}

\item{cluster}{Number of clusters.}

\item{dataset}{Name of dataset.}

\item{timeInternal}{Array with the internal validation calculation times of
the clustering.}

\item{dunn}{Array with the calculation of the dunn for each of the variables.}

\item{connectivity}{Array with the calculation of the connectivity for each
of the variables.}

\item{silhouette}{Array with the calculation of the silhouette for each of
the variables.}

\item{variables}{True if we want to show the value of the metric calculation
and false if we want to show the variable.}
}
\value{
Returns an array with the calculation of each metric based on the
indicated position.
}
\description{
Method that returns the value or variable depending on where it is in the
calculated metrics.
}
\keyword{internal}
