% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calculate_validation_external_by_metrics}
\alias{calculate_validation_external_by_metrics}
\title{Method that calculates which algorithm behaves best for the datasets provided.}
\usage{
calculate_validation_external_by_metrics(df)
}
\arguments{
\item{df}{Data matrix or data frame.}
}
\value{
Return a table with the best performing algorithm for the provided
datasets.
}
\description{
Method that calculates which algorithm behaves best for the datasets provided.
}
\keyword{internal}
