% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{calculate_validation_internal_by_metrics}
\alias{calculate_validation_internal_by_metrics}
\title{Method that calculates which algorithm behaves best for the datasets provided}
\usage{
calculate_validation_internal_by_metrics(df)
}
\arguments{
\item{df}{data matrix or data frame}
}
\value{
returns a table with the best performing algorithm for the provided
datasets
}
\description{
Method that calculates which algorithm behaves best for the datasets provided
}
\keyword{internal}
