% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot.R
\name{cluster_plot}
\alias{cluster_plot}
\title{Plotfunction for Clusters}
\usage{
cluster_plot(df, cluster_center = NULL, nclusters = 2,
  g_mixture = FALSE, theta_comp = c(0.5, 0, 0.1, 1, 0.1),
  theta_cor = NULL, x_label = "X-Achse", y_label = "Y-Achse")
}
\arguments{
\item{df}{The daraframe containing the Clusters. Each column is
interpreted as one point an the colname is the name of the
Cluster the column belongs to.}

\item{cluster_center}{A dataframe of the same configuration as
the data depicting the compute clustercenters.}

\item{nclusters}{The number of clusters to be plotted.}

\item{g_mixture}{A logical vlaue. "TRUE" only used for two
component gaussian mixture.}

\item{theta_comp}{The values of the computed gaussian mixture
in the following order: Mixing ratio, expected value of the
first cluster, the belonging standard deviation, the expected
value pf the second cluster and the belonging standard deviation.}

\item{theta_cor}{The correct gaussian mixture in the same order as
\code{theta_comp}.}

\item{x_label}{A custom label for the x axis.}

\item{y_label}{A custom label for the y axis.}
}
\description{
This plot function supports one-dimensional gaussian mixture clustering
as well as multidimensional k-means, k-medoids and agglomerative clustering.
If multidimensional data is given as input the plot projects onto the first
two dimensions.
}
\details{
If g_mixture is FALSE you are able to input custom cluster_centers.
   If you decide not to, the cluster means are computed and plotted.
   If you want to plot a previously computed gaussian mixture g_mixture needs
   to be TRUE. In the plot you will get the responsibiliity curve of your two
   clusters as well as the computed density. Should you know the correct
   means and standard deviations of your data (e.g. if you did simulate them
   yourself) you can plot the correct gaussian mixture as well by setting
   theta_cor on a vector of your computed values.
}
