% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering-package.R
\docType{package}
\name{Clustering-package}
\alias{Clustering}
\alias{Clustering-package}
\title{Clustering: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
