% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validation.R
\name{connectivity}
\alias{connectivity}
\title{Method to calculate the connectivity}
\usage{
connectivity(
  distance = NULL,
  clusters,
  datadf = NULL,
  neighbSize = 10,
  method = "euclidean"
)
}
\arguments{
\item{distance}{dissimilarity matrix}

\item{clusters}{array that containe tha data grouped in cluster}

\item{neighbSize}{number of neighbours}

\item{method}{indicates the method for calculating distance between points. Default is euclidean}

\item{data}{dataframe with original data}
}
\value{
returns a double with the result of the connectivity calculation
}
\description{
Method to calculate the connectivity
}
\keyword{internal}
