% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_toOrdinal}
\alias{convert_toOrdinal}
\title{Method to convert columns to ordinal.}
\usage{
convert_toOrdinal(df)
}
\arguments{
\item{df}{data frame with the results.}
}
\value{
convert data frame to Ordinal.
}
\description{
Method to convert columns to ordinal.
}
