% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data_sim}
\alias{data_sim}
\title{Datasimulation}
\usage{
data_sim(ncluster_points = 10, midpoints = c(1, 1), sd = 0.5,
  dim = 2)
}
\arguments{
\item{ncluster_points}{The number of points simulated in each cluster.
Either a single value or a vector of values.}

\item{midpoints}{The points the cluster should be grouped around.}

\item{sd}{The standard deviations for the clusters.}

\item{dim}{The dimension of the simulated points.}
}
\value{
A dataframe filled with the simulated points.
}
\description{
A simulating algorithm which simulates normal distributed points
along a set of given midpoints.
}
