% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataframe_by_metrics_evaluation}
\alias{dataframe_by_metrics_evaluation}
\title{Method to filter only the external measurement columns}
\usage{
dataframe_by_metrics_evaluation(data, external = T)
}
\arguments{
\item{data}{information matrix.}

\item{external}{boolean indicating whether it is an external measurement.}
}
\value{
returns a data frame with the filtered columns.
}
\description{
Method to filter only the external measurement columns
}
\keyword{internal}
