% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_definition_attribute}
\alias{detect_definition_attribute}
\title{Method in charge of detecting the limit of a dataset header.}
\usage{
detect_definition_attribute(path)
}
\arguments{
\item{path}{of the dataset}
}
\value{
The row where the dataset attributes definition ends
}
\description{
Method in charge of detecting the limit of a dataset header.
}
\keyword{internal}
