% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diana.R
\name{diana_euclidean_method}
\alias{diana_euclidean_method}
\title{Method that runs the diana algorithm using the euclidean metric to make an external or internal validation of the cluster}
\usage{
diana_euclidean_method(data, clusters, metric)
}
\arguments{
\item{data}{matrix or data frame}

\item{clusters}{number of clusters}

\item{metric}{metrics avalaible in the package. The metrics implemented are: entropy, variation_information,precision,recall,f_measure,fowlkes_mallows_index,connectivity,dunn,silhouette.}
}
\value{
returns a list with both the internal and external evaluation of the grouping
}
\description{
Method that runs the diana algorithm using the euclidean metric to make an external or internal validation of the cluster
}
\keyword{internal}
