% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{em_algorithm}
\alias{em_algorithm}
\title{EM-Algorithm}
\usage{
em_algorithm(vec, epsilon = 1e-04)
}
\arguments{
\item{vec}{The data to be clustered.}

\item{epsilon}{A value depicting the accuracy of the algorithm.}
}
\value{
a dataframe of length five which hast the mixing ration at the first place.
Second and Third place holds the first expeted value and its variance.
Fourth and fifth place shows the second expeted value and its variance.
}
\description{
\code{em_algorithm} takes a one dimensional dataset and gives back the variance
  an the expected value for a two component gaussian mixture
}
\section{Warning}{

The Input needs to be one-dimensional.
}

\examples{
testdata0<-c(-0.39,0.12,0.94,1.67,1.76,2.44,3.72,4.28,4.92,5.53,
  0.06,0.48,1.01,1.68,1.8,3.25,4.12,4.6,5.28,6.22)
em_algorithm(testdata0)

testdata1<-c(0.39,0.12,0.14,0.48,0.44,1.01,1,1.59,1.6,1.67,2.44,
  3.2,3.22,3.25,4.12,4.28,4.52,4.6)
em_algorithm(testdata1)

testdata2 <- gaussian_mixture_sim(mixratio = 0.3, mu1 = 0, sigma1 = 0.5,
  mu2 = 3, sigma2 = 0.8, npoints = 200)
theta2 <- em_algorithm(testdata2)

}
