% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_validation.R
\name{entropy_formula}
\alias{entropy_formula}
\title{Method for calculating entropy}
\usage{
entropy_formula(x_vec)
}
\arguments{
\item{x_vec}{with datas to calculate entropy}
}
\value{
an array with the calculate
}
\description{
Method for calculating entropy
}
\keyword{internal}
