% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{evaluate_validation_internal_by_metrics}
\alias{evaluate_validation_internal_by_metrics}
\title{Method that calculates which algorithm behaves best for the datasets provided.}
\usage{
evaluate_validation_internal_by_metrics(df)
}
\arguments{
\item{df}{data matrix or data frame with the result of running the clustering algorithm.}
}
\value{
returns a table with the best performing algorithm for the provided datasets.
}
\description{
Method that calculates which algorithm behaves best for the datasets provided.
}
\examples{

df = clustering(df = cluster::agriculture, min = 4, max = 5, algorithm='gmm', variables = TRUE)

evaluate_validation_internal_by_metrics(df)

df = clustering(df = Clustering::weather, min = 4, max = 5, algorithm = c("gmm", "kmeans_rcpp"), variables = F)

evaluate_validation_internal_by_metrics(df)

}
