% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{export_file_external}
\alias{export_file_external}
\title{Method that exports the results of external measurements in latex format to a file}
\usage{
export_file_external(df, path = NULL)
}
\arguments{
\item{df}{is a dataframe that contains as a parameter a table in latex format with the results of the external validations}

\item{path}{is the path to a directory where a file is to be stored in latex format}
}
\description{
Method that exports the results of external measurements in latex format to a file
}
\examples{
result = clustering(
               df = cluster::agriculture,
               min = 4,
               max = 5,
               algorithm='gmm',
               metrics=c("precision"),
               variables = TRUE
         )

export_file_external(result)
file.remove("external_data.tex")

}
