% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extension_file}
\alias{extension_file}
\title{Method that return the extension of a file}
\usage{
extension_file(path)
}
\arguments{
\item{path}{dataset directory}
}
\value{
return the extension of file
}
\description{
Method that return the extension of a file
}
\keyword{internal}
