% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_cluster_vector}
\alias{fill_cluster_vector}
\title{Method that fill vector}
\usage{
fill_cluster_vector(data, appcluster)
}
\arguments{
\item{data}{matriz or dataframe with dataset}

\item{appcluster}{data with the information of the appcluster object}
}
\value{
a vector fill with information
}
\description{
Method that fill vector
}
\keyword{internal}
