% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_mixture_sim.R
\name{gaussian_mixture_sim}
\alias{gaussian_mixture_sim}
\title{Simulating Gaussian Mixture}
\usage{
gaussian_mixture_sim(mixratio = 0.5, mu1 = 0, sigma1 = 0.5,
  mu2 = 1, sigma2 = 0.5, npoints = 10)
}
\arguments{
\item{mixratio}{The mixing ratio of the gaussian mixture.}

\item{mu1}{The first mean.}

\item{sigma1}{The square of the standard deviation belonging to the first cluster.}

\item{mu2}{The second mean.}

\item{sigma2}{The square of the standard deviation of the second cluster.}

\item{npoints}{the number of points for each cluster that should be generated.}
}
\value{
An atomic vector filled with the simulated data.
}
\description{
A simulating algorithm which simulates points for a given
two-component gaussian mixture.
}
