% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{gmm_manhattan_method}
\alias{gmm_manhattan_method}
\title{Method that runs the GMM algorithm using the Manhattan metric to make an external or internal validation of the cluster}
\usage{
gmm_manhattan_method(data, clusters, columnClass, metric)
}
\arguments{
\item{data}{matrix or data frame}

\item{clusters}{number of clusters}

\item{columnClass}{number of column, for example if a dataset has five column,
we can select column four to calculate alidation}

\item{metric}{metrics avalaible in the package. The metrics implemented are: entropy, variation_information,precision,recall,f_measure,fowlkes_mallows_index,connectivity,dunn,silhouette.}
}
\value{
returns a list with both the internal and external evaluation of the grouping
}
\description{
Method that runs the GMM algorithm using the Manhattan metric to make an external or internal validation of the cluster
}
\keyword{internal}
