% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{information_external}
\alias{information_external}
\title{Method that returns an array with the external information of the cluster}
\usage{
information_external(metrics, information, size, variables)
}
\arguments{
\item{metrics}{array with the metrics used in the execution of the package}

\item{size}{external number of columns}

\item{variables}{Null returns the position of the variable, otherwise it
returns the value of the variable}

\item{info}{list with external clustering information}
}
\value{
array with the information from the calculation of the external
evaluation of the clustering
}
\description{
Method that returns an array with the external information of the cluster
}
\keyword{internal}
