% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{information_internal}
\alias{information_internal}
\title{Method that returns an array with the internal information of the cluster}
\usage{
information_internal(metrics, information, size, variables)
}
\arguments{
\item{metrics}{array with the metrics used in the execution of the package}

\item{information}{list with internal clustering information}

\item{size}{internal number of columns}

\item{variables}{Null returns the position of the variable, otherwise it
returns the value of the variable}
}
\value{
array with the information from the calculation of the internal
evaluation of the clustering
}
\description{
Method that returns an array with the internal information of the cluster
}
\keyword{internal}
