% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validation.R
\name{initializeInternalValidation}
\alias{initializeInternalValidation}
\title{Method that return a list of external validation initialized to zero.}
\usage{
initializeInternalValidation()
}
\value{
A list of all values set to zero.
}
\description{
Method that return a list of external validation initialized to zero.
}
\keyword{internal}
