% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{is_External_Metrics}
\alias{is_External_Metrics}
\title{Method that checks for external metrics}
\usage{
is_External_Metrics(metrics)
}
\arguments{
\item{metrics}{array with the metrics used in the execution of the package}
}
\value{
true if it exists and false otherwise
}
\description{
Method that checks for external metrics
}
\keyword{internal}
