% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{K-Means Algorithm}
\usage{
k_means(data, K, means = gsCenters(K, data), tol = 0.001,
  maxIter = 10)
}
\arguments{
\item{data}{The data to be clustered. It needs to be a Dataframe (or matrix or list), where each column represents a point.}

\item{K}{The number of clusters.}

\item{means}{Initial guesses for the means of the clusters.}

\item{tol}{The tolerance to which to be clustered to with respect to the matrixnorm induced by squared euclidean distance.}

\item{maxIter}{The maximum number of iterations until the algorithm stops.}
}
\value{
A list list("cluster" = ..., "means" = ...) of the clustering and the computed means. Both entries of the output are dataframes
  where a column represents a point and the colname depicts the cluster the point belongs to.
}
\description{
K-Means Algorithm
}
