% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_medoids.R
\name{k_medoids}
\alias{k_medoids}
\title{K-Medoids Algorithm}
\usage{
k_medoids(data, K, medoids = gsCenters(K, data), tol = 0.01,
  maxIter = 10, d = eucl_dist)
}
\arguments{
\item{data}{The data to be clustered. It needs to be a Dataframe (or matrix or list), where each column represents a point.}

\item{K}{The number of clusters.}

\item{medoids}{Initial guesses for the medoids of the clusters. Default: random guess.}

\item{tol}{The tolerance with respect to the matrixnorm defined by the chosen distance until the algorithm terminates.}

\item{maxIter}{The maximum number of iterations until the algorithm stops.}

\item{d}{A function depicting the distance between to points.}
}
\value{
A list of the clustering and the computed medoids. Both entries of the output are dataframes where a column represents a point and the colname depicts the cluster the point belongs to.
}
\description{
K-Medoids Algorithm
}
