% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans.R
\name{Kmeans}
\alias{Kmeans}
\title{Basic K-Means Clustering}
\usage{
Kmeans(X, centers, max_iter = 100, tol = 1e-04)
}
\arguments{
\item{X}{A numeric matrix of data. Each row is an observation, and each
column is a variable.}

\item{centers}{Either the number of clusters (k) or a matrix of initial
cluster centers.}

\item{max_iter}{The maximum number of iterations allowed.}

\item{tol}{The tolerance for convergence. The algorithm stops if the sum of
squared distances between old and new centers is less than this value.}
}
\value{
A list with the following components:
\item{centers}{A matrix of the final cluster centers.}
\item{cluster}{A vector of integers (from 1:k) indicating the cluster
  to which each point is allocated.}
\item{iter}{The number of iterations performed.}
\item{withinss}{A vector of the within-cluster sum of squares for each
  cluster.}
\item{tot.withinss}{The sum of all `withinss`.}
}
\description{
Performs k-means clustering on a numeric matrix using Lloyd's algorithm.
}
\examples{
# Use the iris dataset
X_iris = as.matrix(iris[, 1:4])

# Run k-means with k=3
result = Kmeans(X_iris, centers = 3, max_iter = 100, tol = 1e-4)
print(result$centers)

}
