% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_cpp.R
\name{Kmeans_cpp}
\alias{Kmeans_cpp}
\title{K-Means Clustering (C++ Implementation)}
\usage{
Kmeans_cpp(X, centers, max_iter = 100, tol = 1e-04)
}
\arguments{
\item{X}{A numeric matrix of data.}

\item{centers}{Initial centers matrix.}

\item{max_iter}{Maximum iterations.}

\item{tol}{Tolerance for convergence.}
}
\value{
A list containing clusters and centers.
}
\description{
K-Means Clustering (C++ Implementation)
}
