% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{max_value_metric}
\alias{max_value_metric}
\title{Method that return max value of metric}
\usage{
max_value_metric(df, metric)
}
\arguments{
\item{df}{data matrix or data frame}

\item{metric}{metric to evaluate}
}
\value{
a value with maximum column
}
\description{
Method that return max value of metric
}
\keyword{internal}
