% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{measure_advclust}
\alias{measure_advclust}
\title{Metrics of the advclust algorithm}
\usage{
measure_advclust()
}
\value{
list with the metrics
}
\description{
Metrics of the advclust algorithm
}
\keyword{internal}
