% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{measure_apcluster}
\alias{measure_apcluster}
\title{Metrics of the apcluster algorithm}
\usage{
measure_apcluster()
}
\value{
list with the metrics
}
\description{
Metrics of the apcluster algorithm
}
\keyword{internal}
