% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{measure_calculate}
\alias{measure_calculate}
\title{Method that returns all the measures executed by the package from the
indicated algorithms}
\usage{
measure_calculate(algorithm)
}
\arguments{
\item{algorithm}{algorithms array}
}
\value{
array with the measures we're going to run
}
\description{
Method that returns all the measures executed by the package from the
indicated algorithms
}
\keyword{internal}
