% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{measure_cluster}
\alias{measure_cluster}
\title{Metrics of the cluster algorithm}
\usage{
measure_cluster()
}
\value{
list with the metrics
}
\description{
Metrics of the cluster algorithm
}
\keyword{internal}
