% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{measure_clusterr}
\alias{measure_clusterr}
\title{Metrics of the ClusterR algorithm}
\usage{
measure_clusterr()
}
\value{
list with the metrics
}
\description{
Metrics of the ClusterR algorithm
}
\keyword{internal}
