% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{measure_package}
\alias{measure_package}
\title{Method that returns all the measures executed by the package}
\usage{
measure_package(package)
}
\arguments{
\item{package}{package array}
}
\value{
array with the measures we're going to run
}
\description{
Method that returns all the measures executed by the package
}
\keyword{internal}
