% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{measure_pvclust}
\alias{measure_pvclust}
\title{Metrics of the pvclust algorithm}
\usage{
measure_pvclust()
}
\value{
list with the metrics
}
\description{
Metrics of the pvclust algorithm
}
\keyword{internal}
