% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{metrics_calculate}
\alias{metrics_calculate}
\title{Method in charge of verifying the implemented metrics}
\usage{
metrics_calculate(metrics)
}
\arguments{
\item{metrics}{array with the metrics used in the execution of the package}
}
\value{
list of metrics
}
\description{
Method in charge of verifying the implemented metrics
}
\keyword{internal}
