% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{metrics_external}
\alias{metrics_external}
\title{Method that returns the list of used external metrics}
\usage{
metrics_external()
}
\value{
external metrics listing array
}
\description{
Method that returns the list of used external metrics
}
\keyword{internal}
