% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{metrics_internal}
\alias{metrics_internal}
\title{Method that returns the list of used internal metrics}
\usage{
metrics_internal()
}
\value{
internal metrics listing array
}
\description{
Method that returns the list of used internal metrics
}
\keyword{internal}
