% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{metrics_validate}
\alias{metrics_validate}
\title{Method that returns the list of used metrics}
\usage{
metrics_validate()
}
\value{
metrics listing array
}
\description{
Method that returns the list of used metrics
}
\keyword{internal}
