% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{new_dbscan}
\alias{new_dbscan}
\title{Dbscan algorithm}
\usage{
new_dbscan(data, eps, minPts)
}
\arguments{
\item{data}{A matrix, data.frame, tibble, with two numeric columns and at least two rows}

\item{eps}{A number. Max distance to other points}

\item{minPts}{A number of how many points should a cluster should at least have}
}
\value{
An object of class `dbscan` with components:
\item{data}{original data}
\item{clusterIds}{A numeric vector with the cluster assignments}
\item{clusteredData}{A matrix with the points and its assigned cluster}
}
\description{
Use this algorithm to identify clusters with a certain density.
}
\details{
Clusters are detected if every point have a max distance of \code{eps} to other points and
if a cluster contains at least \code{minPts} points. The other points are noise marked with -1
in the result
}
\examples{
n <- 30
data <- cbind(xval = c(runif(n, min=0, max=3), runif(n, min=5, max=6), runif(n, min=8, max=10)),
              yval = c(runif(n, min=0, max=3), runif(n, min=2, max=7), runif(n, min=6, max=10)))

db <- new_dbscan(data, 0.5, 3)
plot(db, titles = "regwhe5h")
print(db)
}
