% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{new_optics}
\alias{new_optics}
\title{Optics algorithm}
\usage{
new_optics(data, minPts, num_cluster)
}
\arguments{
\item{data}{A matrix, data.frame, tibble, with two numeric colums and at least two rows}

\item{minPts}{A number of how many points should a cluster should at least have}

\item{num_cluster}{A number. How many clusters should be extracted}
}
\value{
An object of class `optics` with components:
\item{raw_data}{original data}
\item{reachability}{the reachability scores}
\item{ordered_data}{the orderd poins by reachability}
\item{clusterIds}{the cluster assignments}
}
\description{
Use this algorithm to identify a given number of clusters. Optics can
identify clusters with different densities unlike Dbscan.
}
\details{
Clusters can be found by looking at valleys in the ordered reachability distances of each data
point. Optics in its original form does not select clusters for the user instead the user must
found cluster by look at the reachability plot and find valleys. However this implentation tries
to find a given number of clusters \code{num_cluster}.
}
\examples{
n <- 100
data <- cbind(xval = c(runif(n*2, min=0, max=3), runif(n, min=5, max=6), runif(n/11, min=8, max=10)),
              yval = c(runif(n*2, min=0, max=3), runif(n, min=2, max=7), runif(n/11, min=6, max=10)))

res <- new_optics(data = data, minPts = 7, num_cluster = 3)
plot(res)
print(res)
}
