% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{number_columnas_internal}
\alias{number_columnas_internal}
\title{Method that returns how many internal metrics there are in the array of
metrics used in the calculation}
\usage{
number_columnas_internal(metrics)
}
\arguments{
\item{metrics}{array with the metrics used in the execution of the package}
}
\value{
returns the number of occurrences
}
\description{
Method that returns how many internal metrics there are in the array of
metrics used in the calculation
}
\keyword{internal}
