% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_variables_dataset}
\alias{number_variables_dataset}
\title{Method that returns the number of variables in a dataset directory}
\usage{
number_variables_dataset(path)
}
\arguments{
\item{path}{dataset directory}
}
\value{
returns the number of variables in a dataset directory
}
\description{
Method that returns the number of variables in a dataset directory
}
\keyword{internal}
