% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{packages}
\alias{packages}
\title{Method that returns the list of used packages}
\usage{
packages()
}
\value{
package listing array
}
\description{
Method that returns the list of used packages
}
\keyword{internal}
