% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_euclidean_method}
\alias{pam_euclidean_method}
\title{Method that runs the pam algorithm using the Euclidean metric to make an external or internal validation of the cluster.}
\usage{
pam_euclidean_method(dt, clusters, columnClass, metric)
}
\arguments{
\item{dt}{matrix or data frame with the set of values to be applied to the algorithm.}

\item{clusters}{number of clusters.}

\item{columnClass}{is an integer with the number of columns, for example if a dataset has five column,
we can select column four to calculate alidation.}

\item{metric}{is a characters vector with the metrics avalaible in the package. The metrics implemented are: entropy, variation_information,precision,recall,f_measure,fowlkes_mallows_index,connectivity,dunn,silhouette.}
}
\value{
returns a list with both the internal and external evaluation of the grouping.
}
\description{
Method that runs the pam algorithm using the Euclidean metric to make an external or internal validation of the cluster.
}
\keyword{internal}
