% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_dataset}
\alias{path_dataset}
\title{Method that return a list of files that exists in a directory}
\usage{
path_dataset(directory)
}
\arguments{
\item{directory}{of the directory}
}
\value{
a vector with the files existing into of a directory
}
\description{
Method that return a list of files that exists in a directory
}
\keyword{internal}
