% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{plot.dbscan}
\alias{plot.dbscan}
\title{Plot dbscan class}
\usage{
\method{plot}{dbscan}(obj, titles = "Cluster Plot", ...)
}
\arguments{
\item{obj}{A dbscan class, call \code{new_dbscan()} to create an dbscan object}

\item{titles}{A character vector with length 1. Here the titles of the plot}
}
\description{
Plot dbscan class
}
\examples{
n <- 30
data <- cbind(xval = c(runif(n, min=0, max=3), runif(n, min=5, max=6), runif(n, min=8, max=10)),
              yval = c(runif(n, min=0, max=3), runif(n, min=2, max=7), runif(n, min=6, max=10)))

db <- new_dbscan(data, 0.5, 3)
plot(db, titles = "regwhe5h")
}
