% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMeans.R
\name{plot.kMeans}
\alias{plot.kMeans}
\title{Plot kMeans class}
\usage{
\method{plot}{kMeans}(obj, titles = "Cluster Plot", ...)
}
\arguments{
\item{obj}{A kMeans class, call \code{new_kMeans()} to create an kMeans object}

\item{titles}{A character vector with length 1. Here the titles of the plot}
}
\description{
Plot kMeans class
}
\examples{
n <- 50
data <- cbind(wrw = c(runif(n, min=0, max=3), runif(n, min=5, max=6), runif(n, min=8, max=10)),
              yys = c(runif(n, min=0, max=3), runif(n, min=2, max=7), runif(n, min=6, max=10)))

km <- new_kMeans(data, 3, 20, type = "kMeans++")


#visualize
plot(km, titles = "ewgfsgw")
}
