% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{plot.optics}
\alias{plot.optics}
\title{Plot optics class}
\usage{
\method{plot}{optics}(
  obj,
  orderline = FALSE,
  type = "reachability_cluster",
  titles = c("Cluster plot", "Reachability plot"),
  cluster_colored = TRUE,
  ...
)
}
\arguments{
\item{obj}{A optics class, call \code{new_optics()} to create an optics object}

\item{orderline}{A logical to determine if the data points should be connected with lines in reachability order}

\item{type}{A character to decide which plot should be displayed: `reachability_cluster` show both plots, `reachability` show reachability plot
`cluster` show cluster plot}

\item{titles}{A character vector with length 2. Here the titles of the plots for reachability and cluster}

\item{cluster_colored}{A logical. Should the data be colored by the cluster assignments}
}
\description{
Plot optics class
}
\examples{
n <- 100
data <- cbind(xval = c(runif(n*2, min=0, max=3), runif(n, min=5, max=6), runif(n/11, min=8, max=10)),
              yval = c(runif(n*2, min=0, max=3), runif(n, min=2, max=7), runif(n/11, min=6, max=10)))

res <- new_optics(data = data, minPts = 7, num_cluster = 3)
plot(res)
print(res)
}
