% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{plot_clustering}
\alias{plot_clustering}
\title{Method that graphically compares external and internal evaluation metrics.}
\usage{
plot_clustering(df, metric)
}
\arguments{
\item{df}{data matrix or data frame with the result of running the clustering algorithm.}

\item{metric}{string with the name of the metric select to evaluate.}
}
\description{
Method that graphically compares external and internal evaluation metrics.
}
\examples{

result = clustering(
               df = cluster::agriculture,
               min = 4,
               max = 5,
               algorithm='gmm',
               metrics=c("precision"),
               variables = TRUE
         )

plot_clustering(result,c("precision"))

}
