% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{plot_external_validation}
\alias{plot_external_validation}
\title{Method that graphically compares external evaluation metrics}
\usage{
plot_external_validation(df, metric)
}
\arguments{
\item{df}{df data matrix or data frame}

\item{metric}{string with the name of the metric select to evaluate}
}
\description{
Method that graphically compares external evaluation metrics
}
\examples{

df <- clustering(df = cluster::agriculture, min = 4, max = 5, algorith='gmm')

plot_external_validation(df,"precision")

}
