% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{print.dbscan}
\alias{print.dbscan}
\title{Print dbscan class}
\usage{
\method{print}{dbscan}(obj, ...)
}
\arguments{
\item{obj}{A optics class, call \code{new_dbscan()} to create an dbscan object}
}
\description{
Print dbscan class
}
\examples{
n <- 30
data <- cbind(xval = c(runif(n, min=0, max=3), runif(n, min=5, max=6), runif(n, min=8, max=10)),
              yval = c(runif(n, min=0, max=3), runif(n, min=2, max=7), runif(n, min=6, max=10)))

db <- new_dbscan(data, 0.5, 3)
print(db)
}
