% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{print.evaluate_best_validation_external_by_metrics}
\alias{print.evaluate_best_validation_external_by_metrics}
\title{Method that shows on screen which algorithm and which metric behaves best for the datasets provided.}
\usage{
\method{print}{evaluate_best_validation_external_by_metrics}(x, ...)
}
\arguments{
\item{x}{list containing the result of the evaluate_best_validation_external_by_metrics run.}

\item{...}{other params.}
}
\description{
Method that shows on screen which algorithm and which metric behaves best for the datasets provided.
}
\keyword{internal}
