% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMeans.R
\name{print.kMeans}
\alias{print.kMeans}
\title{Print kMeans class}
\usage{
\method{print}{kMeans}(obj, ...)
}
\arguments{
\item{obj}{A kMeans class, call \code{new_kMeans()} to create an kMeans object}
}
\description{
Print kMeans class
}
\examples{
n <- 50
data <- cbind(wrw = c(runif(n, min=0, max=3), runif(n, min=5, max=6), runif(n, min=8, max=10)),
              yys = c(runif(n, min=0, max=3), runif(n, min=2, max=7), runif(n, min=6, max=10)))

km <- new_kMeans(data, 3, 20, type = "kMeans++")

print(km)
}
