% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{print.optics}
\alias{print.optics}
\title{Print optics class}
\usage{
\method{print}{optics}(obj, ...)
}
\arguments{
\item{obj}{A optics class, call \code{new_optics()} to create an optics object}
}
\description{
Print optics class
}
\examples{
n <- 100
data <- cbind(xval = c(runif(n*2, min=0, max=3), runif(n, min=5, max=6), runif(n/11, min=8, max=10)),
              yval = c(runif(n*2, min=0, max=3), runif(n, min=2, max=7), runif(n/11, min=6, max=10)))

res <- new_optics(data = data, minPts = 7, num_cluster = 3)
print(res)
}
