% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvpick.R
\name{pvpick_method}
\alias{pvpick_method}
\title{Method that runs the pvpick algorithm using an external or internal validation of the cluster.}
\usage{
pvpick_method(dt, clusters, columnClass, metric)
}
\arguments{
\item{dt}{Matrix or data frame with the set of values to be applied to the
algorithm.}

\item{clusters}{It's an integer that indexes the number of clusters we want to
create.}

\item{metric}{It's a characters vector with the metrics avalaible in the
package. The metrics implemented are: Entropy, Variation_information,
Precision,Recall,F_measure,Fowlkes_mallows_index,Connectivity,Dunn,
Silhouette.}
}
\value{
Return a list with both the internal and external evaluation of the
grouping.
}
\description{
Method that runs the pvpick algorithm using an external or internal validation of the cluster.
}
\keyword{internal}
