% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_arff.R
\name{read_arff}
\alias{read_arff}
\title{This functions below are functions that belongs to the package 'mldr'
 developed by F. Charte and D. Charte.}
\usage{
read_arff(arff_file)
}
\arguments{
\item{arff_file}{Path to the file}
}
\value{
List containing the relation string,
 a named vector for attributes and a data.frame
 for the data section
}
\description{
More information about 'mldr' package is found in this article:
}
\details{
F. Charte, D. Charte. Working with Multilabel Datasets in R: The mldr Package. The R Journal, 2015

Extracts all useful data from an ARFF file in
R objects
}
