% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_file}
\alias{read_file}
\title{Method that converts a dataset into a matrix}
\usage{
read_file(path)
}
\arguments{
\item{path}{dataset directory}
}
\value{
returns a matrix whose content is the dataset received as a parameter
}
\description{
Method that converts a dataset into a matrix
}
\keyword{internal}
