% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refactorName}
\alias{refactorName}
\title{Method for refactoring the distance measurement name.}
\usage{
refactorName(nameMeasure)
}
\arguments{
\item{nameMeasure}{name of the distance measure}
}
\value{
a string with the refactored measure name
}
\description{
Method for refactoring the distance measurement name.
}
\keyword{internal}
