% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{result_internal_algorithm_by_metric}
\alias{result_internal_algorithm_by_metric}
\title{Method that returns a table with the algorithm and the metric indicated as parameters}
\usage{
result_internal_algorithm_by_metric(df, algorithm)
}
\arguments{
\item{df}{data matrix or data frame}

\item{algorithm}{on which we will calorder(v1)culate the results}
}
\value{
returns a table with the algorithm and the metric indicated as parameter
}
\description{
Method that returns a table with the algorithm and the metric indicated as parameters
}
\examples{

df = clustering(df = cluster::agriculture, min = 4, max = 5, algorithm='gmm', variables = TRUE)

result_internal_algorithm_by_metric(df$result,'gmm')

}
