% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resultClustering}
\alias{resultClustering}
\title{Method for filtering clustering results.}
\usage{
resultClustering(result)
}
\arguments{
\item{result}{data.frame with clustering results.}
}
\value{
a matrix with the filtered columns.
}
\description{
Method for filtering clustering results.
}
\keyword{internal}
