% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validation.R
\name{row_name_df_internal}
\alias{row_name_df_internal}
\title{Method in charge of obtaining those metrics that are internal from those
indicated}
\usage{
row_name_df_internal(metrics)
}
\arguments{
\item{metrics}{array with the metrics used in the calculation}
}
\value{
returns an array with the metrics that are internal
}
\description{
Method in charge of obtaining those metrics that are internal from those
indicated
}
\keyword{internal}
