% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{show_result_external_algorithm_by_metric}
\alias{show_result_external_algorithm_by_metric}
\title{Method that returns a table with the algorithm and the metric indicated as
parameters.}
\usage{
show_result_external_algorithm_by_metric(df, metric)
}
\arguments{
\item{df}{Data matrix or data frame.}

\item{metric}{String with the metric.}
}
\value{
Return a table with the algorithm and the metric indicated as
parameter.
}
\description{
Method that returns a table with the algorithm and the metric indicated as
parameters.
}
\keyword{internal}
