% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validation.R
\name{silhouette_metric}
\alias{silhouette_metric}
\title{Method to calculate the silhouette.}
\usage{
silhouette_metric(clusters_vector, distance)
}
\arguments{
\item{clusters_vector}{Array that containe tha data grouped in cluster.}

\item{distance}{Dissimilarity matrix.}
}
\value{
Return a double with the result of the silhouette calculation.
}
\description{
Method to calculate the silhouette.
}
\keyword{internal}
