% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Method that format a number with four digits}
\usage{
specify_decimal(x, k)
}
\arguments{
\item{x}{number}

\item{k}{number of decimals}
}
\value{
a number convert to string with four digits
}
\description{
Method that format a number with four digits
}
\keyword{internal}
