% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{spectral_clustering}
\alias{spectral_clustering}
\title{Implementation of the spectral clustering algorithm.}
\usage{
spectral_clustering(data, num_clusters, dim_k, ...)
}
\arguments{
\item{data}{A matrix with two columns of numeric data.
Each row represents one data point. The columns are the dimensions.}

\item{num_clusters}{The number of clusters, a numeric value larger than one.}

\item{dim_k}{The dimension the data should be projected into.}

\item{cluster_fun}{A cluster function that runs on the vecors retrieved
from the projection into R^\code{dim_k}}
}
\value{
The return type of \code{cluster_fun}
}
\description{
\code{spectral_clustering} is a clustering algorithm that emphasizes the distances
data.
}
