% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-data.R
\docType{data}
\name{stock}
\alias{stock}
\title{The data provided are daily stock prices from January 1988 through October
1991, for ten aerospace companies.}
\format{
A data frame with 950 observations on 10 variables:
\describe{
  The data provided are daily stock prices from January 1988 through October
  1991, for ten aerospace companies.
  \item{Company1}{company1 details}
  \item{Company2}{company2 details}
  \item{Company3}{company3 details}
  \item{Company4}{company4 details}
  \item{Company5}{company5 details}
  \item{Company6}{company6 details}
  \item{Company7}{company7 details}
  \item{Company8}{company8 details}
  \item{Company9}{company9 details}
  \item{Company10}{company10 details}
}
}
\source{
KEEL, <http://www.keel.es/>
}
\usage{
data(stock)
}
\description{
The data provided are daily stock prices from January 1988 through October
1991, for ten aerospace companies.
}
\keyword{datasets}
