% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stulong-data.R
\docType{data}
\name{stulong}
\alias{stulong}
\title{The study was performed at the 2nd Department of Medicine, 1st Faculty of
Medicine of Charles University and Charles University Hospital. The data were
 transferred to electronic form by the European Centre of Medical
 Informatics, Statisticsand Epidemiology of Charles University and
 Academy of Sciences.}
\format{
A data frame with 1417 observations on 5 variables.
\describe{
  The study was performed at the 2nd Department of Medicine, 1st Faculty of
  Medicine of Charles University and Charles University Hospital. The data
  were transferred to electronic form by the European Centre of Medical
  Informatics, Statisticsand Epidemiology of Charles University and Academy
  of Sciences.
  \item{a1}{Height}
  \item{a2}{Weight}
  \item{a3}{Blood pressure I systolic (mm Hg)}
  \item{a4}{Blood pressure I diastolic (mm Hg)}
  \item{a5}{ercentage Cholesterol in mg}
}
}
\source{
KEEL, <http://www.keel.es/>
}
\usage{
data(stulong)
}
\description{
The study was performed at the 2nd Department of Medicine, 1st Faculty of
Medicine of Charles University and Charles University Hospital. The data were
 transferred to electronic form by the European Centre of Medical
 Informatics, Statisticsand Epidemiology of Charles University and
 Academy of Sciences.
}
\keyword{datasets}
