% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{[.clustering}
\alias{[.clustering}
\title{Filter metrics in a \code{clustering} object returning a new
\code{clustering} object.}
\usage{
\method{[}{clustering}(clustering, condition = T)
}
\arguments{
\item{clustering}{The \code{clustering} object to filter.}

\item{condition}{Expression to filter the \code{clustering} object.}
}
\description{
Generates a new filtered \code{clustering} object.
}
\details{
This function allows you to filter the data set for a given
evaluation metric. The evaluation metrics available are:
\code{Algorithm, Distance, Clusters, Data, Var, Time, Entropy,
Variation_information, Precision, Recall, F_measure, Fowlkes_mallows_index,
Connectivity, Dunn, Silhouette and TimeAtt}.
}
\examples{

library(Clustering)

result <- clustering(package = 'clusterr', df = Clustering::basketball,
min=3, max=4)

result[Precision > 0.14 & Dunn > 0.11]

}
