% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_validation.R
\name{transform_dataset}
\alias{transform_dataset}
\title{Method for filtering external columns of a dataset.}
\usage{
transform_dataset(df)
}
\arguments{
\item{df}{Data frame with clustering results.}
}
\value{
Dafa frame filtered with the columns of the external measurements.

Exists internal measure
}
\description{
Method for filtering external columns of a dataset.
}
