% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validation.R
\name{transform_dataset_internal}
\alias{transform_dataset_internal}
\title{Method for filtering internal columns of a dataset.}
\usage{
transform_dataset_internal(df)
}
\arguments{
\item{df}{data frame with clustering results.}
}
\value{
dafa frame filtered with the columns of the internal measurements.

Exists internal measure
}
\description{
Method for filtering internal columns of a dataset.
}
