% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{best_ranked_external_metrics}
\alias{best_ranked_external_metrics}
\title{Method that calculates the best rated external metrics}
\usage{
best_ranked_external_metrics(df)
}
\arguments{
\item{df}{data matrix or data frame}
}
\value{
returns a table with the external metrics that has the best rating
}
\description{
Method that calculates the best rated external metrics
}
\examples{

df = clustering(df = cluster::agriculture, min = 4, max = 5, algorithm='gmm', variables = TRUE)

best_ranked_external_metrics(df$result)

}
