% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\docType{package}
\name{Clustering-package}
\alias{Clustering}
\alias{Clustering-package}
\title{Clustering: Clustering}
\description{
This package includes three different techniques 
    for hard clustering, one for soft clustering, as well as two 
    different methods for data simulation and one plot function to 
    visualize the outputs of the cluster algorithms.
}
\details{
This package includes three different techniques
for multidimensional hard clustering (K-Means, K-Meoids and Agglomerative Clustering),
one for one-dimensional soft clustering (EM-Algorithm), as well as two
different methods for data simulation (one-dimensional and multidimensional) and one plot function to
visualize the outputs of the cluster algorithms.
}
\author{
\strong{Maintainer}: Tom Schmid \email{tom.schmid@edv-ecke.de}

Authors:
\itemize{
  \item Lydia Rached \email{rached@stud.uni-heidelberg.de}
  \item Tassia Heuser \email{t.heuser@stud.uni-heidelberg.de}
}

}
\keyword{internal}
