% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validation.R
\name{connectivity_metric}
\alias{connectivity_metric}
\title{Method to calculate the connectivity.}
\usage{
connectivity_metric(distance, clusters_vector, dt, method)
}
\arguments{
\item{distance}{Dissimilarity matrix.}

\item{clusters_vector}{Array that containe tha data grouped in cluster.}

\item{dt}{Dataframe with original data.}

\item{method}{Indicates the method for calculating distance between points.}
}
\value{
Return a double with the result of the connectivity calculation.
}
\description{
Method to calculate the connectivity.
}
\keyword{internal}
