% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_validation.R
\name{dunn_metric}
\alias{dunn_metric}
\title{Method to calculate the dunn}
\usage{
dunn_metric(dist, clusters_vector, dt, me)
}
\arguments{
\item{clusters_vector}{array that containe tha data grouped in cluster}

\item{dt}{dataframe with original data}

\item{distance}{dissimilarity matrix}

\item{method}{indicates the method for calculating distance between points}
}
\value{
returns a double with the result of the dunn calculation
}
\description{
Method to calculate the dunn
}
\keyword{internal}
