% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_validation.R
\name{entropy_metric}
\alias{entropy_metric}
\title{Method to calculate the entropy.}
\usage{
entropy_metric(conversion_data_frame, table_convert, column_dataset_label)
}
\arguments{
\item{conversion_data_frame}{A double with the result of the entropy
calculation.}

\item{table_convert}{Table conversion (variable - cluster).}

\item{column_dataset_label}{Array with the calculation of the clustering
algorithm.}
}
\value{
Return a double with the result of the entropy calculation.
}
\description{
Method to calculate the entropy.
}
\keyword{internal}
