% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_column_dataset.R
\name{evaluate_all_column_dataset}
\alias{evaluate_all_column_dataset}
\title{Method in charge of calculating the average for all datasets using all the
algorithms defined in the application.}
\usage{
evaluate_all_column_dataset(datas, method, cluster, nameDataset, metrics)
}
\arguments{
\item{method}{described the metrics used by each of the algorithms}

\item{cluster}{number of clusters}

\item{nameDataset}{specify the name of dataset like information}

\item{metrics}{array with internal or external metrics}

\item{data}{is a data frame or matrix}
}
\value{
a list with result of external and internal validation applying on
algorithms
}
\description{
Method in charge of calculating the average for all datasets using all the
algorithms defined in the application.
}
\keyword{internal}
