% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{evaluate_best_validation_internal_by_metrics}
\alias{evaluate_best_validation_internal_by_metrics}
\title{Method that calculates which algorithm and which metric behaves best for the datasets provided.}
\usage{
evaluate_best_validation_internal_by_metrics(df)
}
\arguments{
\item{df}{data matrix or data frame with the result of running the clustering algorithm.}
}
\value{
returns a table with the algorithm and the best performing metric for the datasets.
}
\description{
Method that calculates which algorithm and which metric behaves best for the datasets provided.
}
\examples{

result = clustering(df = cluster::agriculture, min = 4, max = 5, algorithm='gmm', variables = TRUE)

evaluate_best_validation_internal_by_metrics(result)

result = clustering(
               df = cluster::agriculture,
               min = 4,
               max = 5,
               algorithm='gmm',
               metrics=c("connectivity"),
               variables = TRUE
         )

evaluate_best_validation_internal_by_metrics(result)
}
