% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{evaluate_validation_external_by_metrics}
\alias{evaluate_validation_external_by_metrics}
\title{Method that calculates which algorithm behaves best for the datasets provided.}
\usage{
evaluate_validation_external_by_metrics(df)
}
\arguments{
\item{df}{data matrix or data frame with the result of running the clustering algorithm.}
}
\value{
returns a table with the best performing algorithm for the provided datasets.
}
\description{
Method that calculates which algorithm behaves best for the datasets provided.
}
\examples{

result = clustering(
               df = cluster::agriculture,
               min = 4,
               max = 5,
               algorithm='kmeans_arma',
               metrics=c("precision"),
               variables = TRUE
         )

evaluate_validation_external_by_metrics(result)

}
