% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{execute_datasets}
\alias{execute_datasets}
\title{Evaluation clustering algorithm.}
\usage{
execute_datasets(
  path,
  df,
  packages,
  algorithm,
  cluster_min,
  cluster_max,
  metrics,
  attributes,
  name_dataframe
)
}
\arguments{
\item{path}{Path where the datasets are located.}

\item{df}{Data matrix or data frame, or dissimilarity matrix, depending on
the value of the argument.}

\item{packages}{Array defining the clustering package. The seven packages
implemented are: cluster, ClusterR, amap, apcluster, pvclust.
By default runs all packages.}

\item{algorithm}{Array with the algorithms that implement the package.
The algorithms implemented are: hclust,apclusterK,
agnes,clara,daisy,diana,fanny,mona,pam,gmm,kmeans_arma,kmeans_rcpp,
mini_kmeans, pvclust.}

\item{cluster_min}{Minimum number of clusters. at least one must be.}

\item{cluster_max}{Maximum number of clusters. cluster_max must be greater or
equal cluster_min.}

\item{metrics}{Array defining the metrics avalaible in the package. The night
metrics implemented are: Entropy, Variation_information, Precision, Recall,
F_measure, Fowlkes_mallows_index, Connectivity, Dunn and Silhouette.}

\item{name_dataframe}{Name of data.frame when df is fill.}
}
\value{
Returns a matrix with the result of running all the metrics of the
algorithms contained in the packages we indicated.
}
\description{
Method of performing information processing
}
\keyword{internal}
