% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{export_file_internal}
\alias{export_file_internal}
\title{Method that exports the results of internal measurements in latex format to a file}
\usage{
export_file_internal(df, path)
}
\arguments{
\item{df}{is a dataframe that contains as a parameter a table in latex format with the results of the internal validations}

\item{path}{is the path to a directory where a file is to be stored in latex format}
}
\description{
Method that exports the results of internal measurements in latex format to a file
}
