% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmed.R
\name{kmedoid}
\alias{kmedoid}
\title{Implementation of a k-medoid clustering algorithm}
\usage{
kmedoid(data, k = 1, iter = 1)
}
\arguments{
\item{data}{A list with two columns of numeric data.}

\item{k}{The number of medoids or clusters to be computed.}

\item{iter}{The number of iterations the optimisation should go through. If an
optimal state has been reached, the function will exit.}
}
\description{
\code{kmedoid()} utilizes a PAM algorithm (partitioning around medoids) to cluster
data. k-medoids aim to minimize the total sum of distances within each cluster
to its central medoid.
}
\examples{
kmedoid(iris[3:4],3,5)
}
