% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{max_value_metric}
\alias{max_value_metric}
\title{Method that return max value of metric.}
\usage{
max_value_metric(df, metric, isExternalMetric)
}
\arguments{
\item{df}{Data matrix or data frame.}

\item{metric}{Metric to evaluate.}
}
\value{
A value with maximum column.
}
\description{
Method that return max value of metric.
}
\keyword{internal}
