% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{show_result_external_algorithm_by_metric}
\alias{show_result_external_algorithm_by_metric}
\title{Method that returns a table with the algorithm and the metric indicated as
parameters}
\usage{
show_result_external_algorithm_by_metric(df, algorithm)
}
\arguments{
\item{df}{data matrix or data frame}

\item{algorithm}{on which we will calculate the results}
}
\value{
returns a table with the algorithm and the metric indicated as
parameter
}
\description{
Method that returns a table with the algorithm and the metric indicated as
parameters
}
\keyword{internal}
