% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{[.clustering}
\alias{[.clustering}
\title{Filter metrics in a \code{clustering} object returning a new \code{clustering} object.}
\usage{
\method{[}{clustering}(clusteringObject, condition = T)
}
\arguments{
\item{clusteringObject}{The \code{clusteringObject} object to filter.}

\item{condition}{Expression to filter the \code{clustering} object}
}
\description{
Generates a new \code{clustering} object containing the metrics that passed the filter
specified.
}
\details{
This function allows you to filter the data set for a given evaluation metric.
The evaluation metrics available are:
\code{Algorithm, Distance, Clusters, Dataset, Ranking, timeExternal, entropy, variation_information, precision, recall, f_measure, fowlkes_mallows_index, connectivity, dunn, silhouette, timeInternal}.
}
\examples{

library(clustering)

datasetTest[precision > 0.03]

Also combine multiple variables:

datasetTest[precision > 0.03 & dunn > 0.1]

}
